Use big integer for users

+++ api/app/data/models/users.py
@@ -12,7 +12,7 @@ class User(Base):
     __tablename__ = 'users'
 
     id = sa.Column(sa.Integer, primary_key=True)
-    telegram_user_id = sa.Column(sa.Integer)
+    telegram_user_id = sa.Column(sa.BigInteger)
     name = sa.Column(sa.String(128))
     phone = sa.Column(sa.String(12))
     created_at = sa.Column(sa.DateTime, default=datetime.now)
